.
Начал прописывать service, добавил по надобности конструкторы в Entity и POJO и изменил repository

+++ logical/src/main/java/com/example/logical/entity/Test.java
@@ -14,7 +14,7 @@ public class Test {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "test_seq")
     @SequenceGenerator(name = "test_seq", sequenceName = "test_seq", allocationSize = 1)
-    private int test_id;
+    private Long test_id;
 
     private String test_name;
